Test.Consensus.Shelley.Examples

Setup

testEpochInfo ∷ EpochInfo Identity

testShelleyGenesis ∷ ShelleyGenesis era

These are dummy values.

codecConfig ∷ CodecConfig (ShelleyBlock StandardShelley)

mkDummyHash ∷ ∀ h a. HashAlgorithm h ⇒ Proxy h → Int → Hash h a

keyToCredential ∷ ShelleyBasedEra era ⇒ KeyPair r (EraCrypto era) → Credential r era

Examples

examplesShelley ∷ Examples (ShelleyBlock StandardShelley)

examplesAllegra ∷ Examples (ShelleyBlock StandardAllegra)

examplesMary ∷ Examples (ShelleyBlock StandardMary)

Era-specific examples

exampleCoin ∷ Coin

Individual examples

exampleBlock ∷ ∀ era. ShelleyBasedEra era ⇒ TxBody era → ShelleyBlock era

This is not a valid block. We don't care, we are only interested in serialisation, not validation.

exampleHeaderHash ∷ ∀ era. ShelleyBasedEra era ⇒ Proxy era → HeaderHash (ShelleyBlock era)

exampleTx ∷ ∀ era. ShelleyBasedEra era ⇒ TxBody era → Tx era

This is not a valid transaction. We don't care, we are only interested in serialisation, not validation.

exampleApplyTxErr ∷ ShelleyBasedEra era ⇒ ApplyTxErr (ShelleyBlock era)

exampleChainDepState ∷ ∀ c. PraosCrypto c ⇒ ChainDepState (TPraos c)

exampleLedgerState ∷ ∀ era. ShelleyBasedEra era ⇒ Value era → LedgerState (ShelleyBlock era)

exampleHeaderState ∷ ShelleyBasedEra era ⇒ HeaderState (ShelleyBlock era)

exampleExtLedgerState ∷ ShelleyBasedEra era ⇒ Value era → ExtLedgerState (ShelleyBlock era)

Keys

examplePayKey ∷ Crypto c ⇒ KeyPair 'Payment c

exampleStakeKey ∷ Crypto c ⇒ KeyPair 'Staking c

exampleKeys ∷ ∀ c r. Crypto c ⇒ AllIssuerKeys c r

examplePoolParams ∷ ∀ era. ShelleyBasedEra era ⇒ PoolParams era