Test.Consensus.Shelley.Generators

Documentation

data SomeResult blk where

To easily generate all the possible results of the Query GADT, we introduce an existential that also bundles the corresponding Query as evidence. We also capture Eq, Show, and Typeable constraints, as we need them in the tests.

Constructors

SomeResult ∷ ∀ blk result. (Eq result, Show result, Typeable result) ⇒ Query blk result → result → SomeResult blk

Orphan instances

Arbitrary ShelleyNodeToNodeVersion

Instance details

Methods

arbitrary ∷ Gen ShelleyNodeToNodeVersion

shrink ∷ ShelleyNodeToNodeVersion → [ShelleyNodeToNodeVersion]

Arbitrary ShelleyNodeToClientVersion

Instance details

Methods

arbitrary ∷ Gen ShelleyNodeToClientVersion

shrink ∷ ShelleyNodeToClientVersion → [ShelleyNodeToClientVersion]

HashAlgorithm h ⇒ Arbitrary (AnnTip (Block h))

Instance details

Methods

arbitrary ∷ Gen (AnnTip (Block h))

shrink ∷ AnnTip (Block h) → [AnnTip (Block h)]

HashAlgorithm h ⇒ Arbitrary (GenTx (Block h))

Instance details

Methods

arbitrary ∷ Gen (GenTx (Block h))

shrink ∷ GenTx (Block h) → [GenTx (Block h)]

HashAlgorithm h ⇒ Arbitrary (GenTxId (Block h))

Instance details

Methods

arbitrary ∷ Gen (GenTxId (Block h))

shrink ∷ GenTxId (Block h) → [GenTxId (Block h)]

(HashAlgorithm h, ∀ a. Arbitrary (Hash h a)) ⇒ Arbitrary (LedgerState (Block h))

Instance details

Methods

arbitrary ∷ Gen (LedgerState (Block h))

shrink ∷ LedgerState (Block h) → [LedgerState (Block h)]

HashAlgorithm h ⇒ Arbitrary (Header (Block h))

Instance details

Methods

arbitrary ∷ Gen (Header (Block h))

shrink ∷ Header (Block h) → [Header (Block h)]

HashAlgorithm h ⇒ Arbitrary (Point (Block h))

Instance details

Methods

arbitrary ∷ Gen (Point (Block h))

shrink ∷ Point (Block h) → [Point (Block h)]

HashAlgorithm h ⇒ Arbitrary (NonMyopicMemberRewards (TPraosMockCrypto h))

Instance details

Methods

arbitrary ∷ Gen (NonMyopicMemberRewards (TPraosMockCrypto h))

shrink ∷ NonMyopicMemberRewards (TPraosMockCrypto h) → [NonMyopicMemberRewards (TPraosMockCrypto h)]

HashAlgorithm h ⇒ Arbitrary (TPraosState (TPraosMockCrypto h))

Instance details

Methods

arbitrary ∷ Gen (TPraosState (TPraosMockCrypto h))

shrink ∷ TPraosState (TPraosMockCrypto h) → [TPraosState (TPraosMockCrypto h)]

Crypto c ⇒ Arbitrary (ShelleyHash c)

Instance details

Methods

arbitrary ∷ Gen (ShelleyHash c)

shrink ∷ ShelleyHash c → [ShelleyHash c]

HashAlgorithm h ⇒ Arbitrary (ApplyTxError (TPraosMockCrypto h))

Instance details

Methods

arbitrary ∷ Gen (ApplyTxError (TPraosMockCrypto h))

shrink ∷ ApplyTxError (TPraosMockCrypto h) → [ApplyTxError (TPraosMockCrypto h)]

HashAlgorithm h ⇒ Arbitrary (SomeResult (Block h))

Instance details

Methods

arbitrary ∷ Gen (SomeResult (Block h))

shrink ∷ SomeResult (Block h) → [SomeResult (Block h)]

HashAlgorithm h ⇒ Arbitrary (LedgerView (TPraosMockCrypto h))

Instance details

Methods

arbitrary ∷ Gen (LedgerView (TPraosMockCrypto h))

shrink ∷ LedgerView (TPraosMockCrypto h) → [LedgerView (TPraosMockCrypto h)]

HashAlgorithm h ⇒ Arbitrary (ChainDepState (TPraosMockCrypto h))

Instance details

Methods

arbitrary ∷ Gen (ChainDepState (TPraosMockCrypto h))

shrink ∷ ChainDepState (TPraosMockCrypto h) → [ChainDepState (TPraosMockCrypto h)]

Arbitrary (PParams' StrictMaybe)

Instance details

Methods

arbitrary ∷ Gen (PParams' StrictMaybe)

shrink ∷ PParams' StrictMaybe → [PParams' StrictMaybe]

HashAlgorithm h ⇒ Arbitrary (Block h)

Instance details

Methods

arbitrary ∷ Gen (Block h)

shrink ∷ Block h → [Block h]

Arbitrary (SomeBlock (NestedCtxt f) (Block h))

Instance details

Methods

arbitrary ∷ Gen (SomeBlock (NestedCtxt f) (Block h))

shrink ∷ SomeBlock (NestedCtxt f) (Block h) → [SomeBlock (NestedCtxt f) (Block h)]

HashAlgorithm h ⇒ Arbitrary (SomeBlock Query (Block h))

Instance details

Methods

arbitrary ∷ Gen (SomeBlock Query (Block h))

shrink ∷ SomeBlock Query (Block h) → [SomeBlock Query (Block h)]

Arbitrary a ⇒ Arbitrary (WithVersion ShelleyNodeToNodeVersion a)
We only have single version, so no special casing required.

This blanket orphan instance will have to be replaced with more specific ones, once we introduce a different Shelley version.

Instance details

Methods

arbitrary ∷ Gen (WithVersion ShelleyNodeToNodeVersion a)

shrink ∷ WithVersion ShelleyNodeToNodeVersion a → [WithVersion ShelleyNodeToNodeVersion a]

Arbitrary a ⇒ Arbitrary (WithVersion ShelleyNodeToClientVersion a)
We only have single version, so no special casing required.

This blanket orphan instance will have to be replaced with more specific ones, once we introduce a different Shelley version.

Instance details

Methods

arbitrary ∷ Gen (WithVersion ShelleyNodeToClientVersion a)

shrink ∷ WithVersion ShelleyNodeToClientVersion a → [WithVersion ShelleyNodeToClientVersion a]