Test.Consensus.Shelley.MockCrypto

Documentation

data TPraosMockCrypto h

A mock replacement for TPraosStandardCrypto

We run the tests with this mock crypto, as it is easier to generate and debug things. The code is parametric in the crypto, so it shouldn't make much of a difference. This also has the important advantage that we can reuse the generators from cardano-ledger-specs.

type Block h = ShelleyBlock (TPraosMockCrypto h)