Test.ThreadNet.Infra.Shelley

Documentation

data CoreNode era

Constructors

CoreNode

Fields

cnGenesisKey ∷ !(SignKeyDSIGN era)

cnDelegateKey ∷ !(SignKeyDSIGN era)

Cold delegate key. The hash of the corresponding verification (public) key will be used as the payment credential.

cnStakingKey ∷ !(SignKeyDSIGN era)

The hash of the corresponding verification (public) key will be used as the staking credential.

cnVRF ∷ !(SignKeyVRF era)

cnKES ∷ !(SignKeyKES era)

cnOCert ∷ !(OCert era)

data CoreNodeKeyInfo era

Constructors

CoreNodeKeyInfo

Fields

cnkiKeyPair ∷ (KeyPair Payment era, KeyPair Staking era)

cnkiCoreNode ∷ (KeyPair Genesis era, AllIssuerKeys era GenesisDelegate)

newtype DecentralizationParam

A suitable value for the d protocol parameter

In the range 0 to 1, inclusive. Beware the misnomer: 0 means fully decentralized, and 1 means fully centralized.

Constructors

DecentralizationParam

Fields

decentralizationParamToRational ∷ Rational

data KesConfig

Currently maxEvolutions * slotsPerEvolution is the max number of slots the test can run without needing new ocerts.

TODO This limitation may be lifted by PR #2107, see https://github.com/input-output-hk/ouroboros-network/issues/2107.

Constructors

KesConfig

Fields

maxEvolutions ∷ Word64

slotsPerEvolution ∷ Word64

coreNodeKeys ∷ Mock (Crypto era) ⇒ CoreNode era → CoreNodeKeyInfo era

genCoreNode ∷ ∀ era. TPraosCrypto era ⇒ KESPeriod → Gen (CoreNode era)

incrementMinorProtVer ∷ ProtVer → ProtVer

mkCredential ∷ TPraosCrypto era ⇒ SignKeyDSIGN era → Credential r era

mkEpochSize ∷ SecurityParam → Rational → EpochSize

The epoch size, given k and f.

INVARIANT: 10 * k / f must be a whole number.

mkGenesisConfig

Arguments

∷ TPraosCrypto era

⇒ ProtVer
Initial protocol version

→ SecurityParam

→ Rational
Initial active slot coefficient

→ DecentralizationParam

→ SlotLength

→ KesConfig

→ [CoreNode era]

→ ShelleyGenesis era

Note: a KES algorithm supports a particular max number of KES evolutions, but we can configure a potentially lower maximum for the ledger, that's why we take it as an argument.

mkKesConfig ∷ ∀ proxy era. Era era ⇒ proxy era → NumSlots → KesConfig

A KesConfig that will not require more evolutions than this test's crypto allows.

mkKeyHash ∷ TPraosCrypto era ⇒ SignKeyDSIGN era → KeyHash r era

mkKeyHashVrf ∷ (HashAlgorithm h, VRFAlgorithm vrf) ⇒ SignKeyVRF vrf → Hash h (VerKeyVRF vrf)

mkLeaderCredentials ∷ TPraosCrypto era ⇒ CoreNode era → TPraosLeaderCredentials era

mkProtocolRealTPraos ∷ ∀ m era. (IOLike m, TPraosCrypto era) ⇒ ShelleyGenesis era → Nonce → ProtVer → CoreNode era → ProtocolInfo m (ShelleyBlock era)

mkSetDecentralizationParamTxs

Arguments

∷ TPraosCrypto era

⇒ [CoreNode era]

→ ProtVer
The proposed protocol version

→ SlotNo
The TTL

→ DecentralizationParam
The new value

→ [GenTx (ShelleyBlock era)]

mkVerKey ∷ TPraosCrypto era ⇒ SignKeyDSIGN era → VKey r era

networkId ∷ Network

tpraosSlotLength ∷ SlotLength