Test.Consensus.Shelley.Examples

Setup

codecConfig ∷ CodecConfig (ShelleyBlock StandardShelley)

keyToCredential ∷ Crypto c ⇒ KeyPair r c → Credential r c

mkDummyHash ∷ ∀ h a. HashAlgorithm h ⇒ Proxy h → Int → Hash h a

testEpochInfo ∷ EpochInfo Identity

testShelleyGenesis ∷ ShelleyGenesis era

These are dummy values.

Examples

examplesAllegra ∷ Examples (ShelleyBlock StandardAllegra)

examplesAlonzo ∷ Examples (ShelleyBlock StandardAlonzo)

examplesMary ∷ Examples (ShelleyBlock StandardMary)

examplesShelley ∷ Examples (ShelleyBlock StandardShelley)

Era-specific examples

exampleCoin ∷ Coin

Individual examples

exampleApplyTxErr ∷ ∀ era. (ShelleyBasedEra era, PredicateFailure (EraRule "DELEGS" era) ~ DelegsPredicateFailure era) ⇒ ApplyTxErr (ShelleyBlock era)

exampleBlock ∷ ∀ era. ShelleyBasedEra era ⇒ TxInBlock era → ShelleyBlock era

This is not a valid block. We don't care, we are only interested in serialisation, not validation.

exampleChainDepState ∷ ∀ c. PraosCrypto c ⇒ ChainDepState (TPraos c)

exampleExtLedgerState ∷ (ShelleyBasedEra era, PParams era ~ PParams era) ⇒ Value era → ExtLedgerState (ShelleyBlock era)

exampleHeaderHash ∷ ∀ era. ShelleyBasedEra era ⇒ Proxy era → HeaderHash (ShelleyBlock era)

exampleHeaderState ∷ ShelleyBasedEra era ⇒ HeaderState (ShelleyBlock era)

exampleLedgerState ∷ ∀ era. (ShelleyBasedEra era, PParams era ~ PParams era) ⇒ Value era → LedgerState (ShelleyBlock era)

exampleTx ∷ ∀ era. ShelleyBasedEra era ⇒ (TxBody era → KeyPairWits era → Witnesses era) → TxBody era → AuxiliaryData era → Tx era

This is not a valid transaction. We don't care, we are only interested in serialisation, not validation.

Keys

exampleKeys ∷ ∀ c r. Crypto c ⇒ AllIssuerKeys c r

examplePayKey ∷ Crypto c ⇒ KeyPair 'Payment c

examplePoolParams ∷ ∀ c. Crypto c ⇒ PoolParams c

exampleStakeKey ∷ Crypto c ⇒ KeyPair 'Staking c