Test.Consensus.Shelley.Examples

Setup

testEpochInfo ∷ EpochInfo Identity

testShelleyGenesis ∷ ShelleyGenesis era

These are dummy values.

codecConfig ∷ CodecConfig (ShelleyBlock StandardShelley)

mkDummyHash ∷ ∀ h a. HashAlgorithm h ⇒ Proxy h → Int → Hash h a

Examples

examples ∷ Examples (ShelleyBlock StandardShelley)

exampleBlock ∷ ShelleyBlock StandardShelley

This is not a valid block. We don't care, we are only interested in serialisation, not validation.

exampleHeader ∷ Header (ShelleyBlock StandardShelley)

exampleHeaderHash ∷ HeaderHash (ShelleyBlock StandardShelley)

exampleTx ∷ ∀ era. era ~ StandardShelley ⇒ Tx era

This is not a valid transaction. We don't care, we are only interested in serialisation, not validation.

exampleGenTx ∷ GenTx (ShelleyBlock StandardShelley)

exampleGenTxId ∷ GenTxId (ShelleyBlock StandardShelley)

exampleApplyTxErr ∷ ApplyTxErr (ShelleyBlock StandardShelley)

exampleChainDepState ∷ ChainDepState (BlockProtocol (ShelleyBlock StandardShelley))

exampleLedgerState ∷ LedgerState (ShelleyBlock StandardShelley)

exampleHeaderState ∷ HeaderState (ShelleyBlock StandardShelley)

exampleExtLedgerState ∷ ExtLedgerState (ShelleyBlock StandardShelley)