Test.Consensus.Shelley.Examples

Setup

type Block = ShelleyBlock TPraosMockCrypto

We run the tests with mock crypto, as it is easier to generate and debug things. The code is parametric in the crypto, so it shouldn't make much of a difference. This also has the important advantage that we can reuse the generators from cardano-ledger-specs.

testEpochInfo ∷ EpochInfo Identity

testShelleyGenesis ∷ ShelleyGenesis c

These are dummy values.

mkDummyHash ∷ ∀ c a. Crypto c ⇒ Proxy c → Int → Hash c a

Examples

exampleBlock ∷ Block

exampleHeader ∷ Header Block

exampleHeaderHash ∷ HeaderHash Block

exampleGenTx ∷ GenTx Block

exampleGenTxId ∷ GenTxId Block

exampleApplyTxErr ∷ ApplyTxErr Block

exampleConsensusState ∷ ConsensusState (BlockProtocol Block)

exampleLedgerState ∷ LedgerState Block

exampleHeaderState ∷ HeaderState Block

exampleExtLedgerState ∷ ExtLedgerState Block