Test.Consensus.Shelley.Generators

Documentation

data SomeResult blk where

To easily generate all the possible results of the Query GADT, we introduce an existential that also bundles the corresponding Query as evidence. We also capture Eq, Show, and Typeable constraints, as we need them in the tests.

Constructors

SomeResult ∷ ∀ blk result. (Eq result, Show result, Typeable result) ⇒ Query blk result → result → SomeResult blk

Orphan instances

Arbitrary ShelleyTransition

Instance details

Methods

arbitrary ∷ Gen ShelleyTransition

shrink ∷ ShelleyTransition → [ShelleyTransition]

Arbitrary ShelleyNodeToNodeVersion

Instance details

Methods

arbitrary ∷ Gen ShelleyNodeToNodeVersion

shrink ∷ ShelleyNodeToNodeVersion → [ShelleyNodeToNodeVersion]

Arbitrary ShelleyNodeToClientVersion

Instance details

Methods

arbitrary ∷ Gen ShelleyNodeToClientVersion

shrink ∷ ShelleyNodeToClientVersion → [ShelleyNodeToClientVersion]

CanMock era ⇒ Arbitrary (AnnTip (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (AnnTip (ShelleyBlock era))

shrink ∷ AnnTip (ShelleyBlock era) → [AnnTip (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (GenTx (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (GenTx (ShelleyBlock era))

shrink ∷ GenTx (ShelleyBlock era) → [GenTx (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (GenTxId (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (GenTxId (ShelleyBlock era))

shrink ∷ GenTxId (ShelleyBlock era) → [GenTxId (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (LedgerState (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (LedgerState (ShelleyBlock era))

shrink ∷ LedgerState (ShelleyBlock era) → [LedgerState (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (Header (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (Header (ShelleyBlock era))

shrink ∷ Header (ShelleyBlock era) → [Header (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (Point (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (Point (ShelleyBlock era))

shrink ∷ Point (ShelleyBlock era) → [Point (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (ShelleyTip era)

Instance details

Methods

arbitrary ∷ Gen (ShelleyTip era)

shrink ∷ ShelleyTip era → [ShelleyTip era]

CanMock era ⇒ Arbitrary (NonMyopicMemberRewards era)

Instance details

Methods

arbitrary ∷ Gen (NonMyopicMemberRewards era)

shrink ∷ NonMyopicMemberRewards era → [NonMyopicMemberRewards era]

Era era ⇒ Arbitrary (TPraosState era)

Instance details

Methods

arbitrary ∷ Gen (TPraosState era)

shrink ∷ TPraosState era → [TPraosState era]

CanMock era ⇒ Arbitrary (ShelleyHash era)

Instance details

Methods

arbitrary ∷ Gen (ShelleyHash era)

shrink ∷ ShelleyHash era → [ShelleyHash era]

CanMock era ⇒ Arbitrary (ShelleyBlock era)

Instance details

Methods

arbitrary ∷ Gen (ShelleyBlock era)

shrink ∷ ShelleyBlock era → [ShelleyBlock era]

CanMock era ⇒ Arbitrary (ApplyTxError era)

Instance details

Methods

arbitrary ∷ Gen (ApplyTxError era)

shrink ∷ ApplyTxError era → [ApplyTxError era]

CanMock era ⇒ Arbitrary (SomeResult (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (SomeResult (ShelleyBlock era))

shrink ∷ SomeResult (ShelleyBlock era) → [SomeResult (ShelleyBlock era)]

(TPraosCrypto era, CanMock era) ⇒ Arbitrary (LedgerView era)

Instance details

Methods

arbitrary ∷ Gen (LedgerView era)

shrink ∷ LedgerView era → [LedgerView era]

Era era ⇒ Arbitrary (ChainDepState era)

Instance details

Methods

arbitrary ∷ Gen (ChainDepState era)

shrink ∷ ChainDepState era → [ChainDepState era]

Era era ⇒ Arbitrary (SomeBlock (NestedCtxt f) (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (SomeBlock (NestedCtxt f) (ShelleyBlock era))

shrink ∷ SomeBlock (NestedCtxt f) (ShelleyBlock era) → [SomeBlock (NestedCtxt f) (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (SomeBlock Query (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (SomeBlock Query (ShelleyBlock era))

shrink ∷ SomeBlock Query (ShelleyBlock era) → [SomeBlock Query (ShelleyBlock era)]

Arbitrary a ⇒ Arbitrary (WithVersion ShelleyNodeToNodeVersion a)
We only have single version, so no special casing required.

This blanket orphan instance will have to be replaced with more specific ones, once we introduce a different Shelley version.

Instance details

Methods

arbitrary ∷ Gen (WithVersion ShelleyNodeToNodeVersion a)

shrink ∷ WithVersion ShelleyNodeToNodeVersion a → [WithVersion ShelleyNodeToNodeVersion a]

Arbitrary a ⇒ Arbitrary (WithVersion ShelleyNodeToClientVersion a)
We only have single version, so no special casing required.

This blanket orphan instance will have to be replaced with more specific ones, once we introduce a different Shelley version.

Instance details

Methods

arbitrary ∷ Gen (WithVersion ShelleyNodeToClientVersion a)

shrink ∷ WithVersion ShelleyNodeToClientVersion a → [WithVersion ShelleyNodeToClientVersion a]

Arbitrary (PParams' StrictMaybe era)

Instance details

Methods

arbitrary ∷ Gen (PParams' StrictMaybe era)

shrink ∷ PParams' StrictMaybe era → [PParams' StrictMaybe era]