Test.Consensus.Shelley.Generators

Documentation

data SomeResult blk where

To easily generate all the possible results of the Query GADT, we introduce an existential that also bundles the corresponding Query as evidence. We also capture Eq, Show, and Typeable constraints, as we need them in the tests.

Constructors

SomeResult ∷ ∀ result blk. (Eq result, Show result, Typeable result) ⇒ BlockQuery blk result → result → SomeResult blk

Orphan instances

Arbitrary ShelleyTransition

Instance details

Methods

arbitrary ∷ Gen ShelleyTransition

shrink ∷ ShelleyTransition → [ShelleyTransition]

Arbitrary ShelleyNodeToNodeVersion

Instance details

Methods

arbitrary ∷ Gen ShelleyNodeToNodeVersion

shrink ∷ ShelleyNodeToNodeVersion → [ShelleyNodeToNodeVersion]

Arbitrary ShelleyNodeToClientVersion

Instance details

Methods

arbitrary ∷ Gen ShelleyNodeToClientVersion

shrink ∷ ShelleyNodeToClientVersion → [ShelleyNodeToClientVersion]

CanMock era ⇒ Arbitrary (AnnTip (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (AnnTip (ShelleyBlock era))

shrink ∷ AnnTip (ShelleyBlock era) → [AnnTip (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (GenTx (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (GenTx (ShelleyBlock era))

shrink ∷ GenTx (ShelleyBlock era) → [GenTx (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (GenTxId (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (GenTxId (ShelleyBlock era))

shrink ∷ GenTxId (ShelleyBlock era) → [GenTxId (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (LedgerState (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (LedgerState (ShelleyBlock era))

shrink ∷ LedgerState (ShelleyBlock era) → [LedgerState (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (Header (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (Header (ShelleyBlock era))

shrink ∷ Header (ShelleyBlock era) → [Header (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (Point (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (Point (ShelleyBlock era))

shrink ∷ Point (ShelleyBlock era) → [Point (ShelleyBlock era)]

PraosCrypto c ⇒ Arbitrary (TPraosState c)

Instance details

Methods

arbitrary ∷ Gen (TPraosState c)

shrink ∷ TPraosState c → [TPraosState c]

PraosCrypto c ⇒ Arbitrary (NonMyopicMemberRewards c)

Instance details

Methods

arbitrary ∷ Gen (NonMyopicMemberRewards c)

shrink ∷ NonMyopicMemberRewards c → [NonMyopicMemberRewards c]

CanMock era ⇒ Arbitrary (ShelleyTip era)

Instance details

Methods

arbitrary ∷ Gen (ShelleyTip era)

shrink ∷ ShelleyTip era → [ShelleyTip era]

Mock c ⇒ Arbitrary (ShelleyHash c)

Instance details

Methods

arbitrary ∷ Gen (ShelleyHash c)

shrink ∷ ShelleyHash c → [ShelleyHash c]

CanMock era ⇒ Arbitrary (ShelleyBlock era)
The upstream Arbitrary instance for Shelley blocks does not generate coherent blocks, so neither does this.

Instance details

Methods

arbitrary ∷ Gen (ShelleyBlock era)

shrink ∷ ShelleyBlock era → [ShelleyBlock era]

CanMock era ⇒ Arbitrary (Coherent (ShelleyBlock era))
This uses a different upstream generator to ensure the header and block body relate as expected.

Instance details

Methods

arbitrary ∷ Gen (Coherent (ShelleyBlock era))

shrink ∷ Coherent (ShelleyBlock era) → [Coherent (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (SomeResult (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (SomeResult (ShelleyBlock era))

shrink ∷ SomeResult (ShelleyBlock era) → [SomeResult (ShelleyBlock era)]

PraosCrypto c ⇒ Arbitrary (ChainDepState c)

Instance details

Methods

arbitrary ∷ Gen (ChainDepState c)

shrink ∷ ChainDepState c → [ChainDepState c]

CanMock era ⇒ Arbitrary (SomeSecond BlockQuery (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (SomeSecond BlockQuery (ShelleyBlock era))

shrink ∷ SomeSecond BlockQuery (ShelleyBlock era) → [SomeSecond BlockQuery (ShelleyBlock era)]

ShelleyBasedEra era ⇒ Arbitrary (SomeSecond (NestedCtxt f) (ShelleyBlock era))

Instance details

Methods

arbitrary ∷ Gen (SomeSecond (NestedCtxt f) (ShelleyBlock era))

shrink ∷ SomeSecond (NestedCtxt f) (ShelleyBlock era) → [SomeSecond (NestedCtxt f) (ShelleyBlock era)]

CanMock era ⇒ Arbitrary (WithVersion ShelleyNodeToClientVersion (SomeSecond BlockQuery (ShelleyBlock era)))
Some Querys are only supported by ShelleyNodeToClientVersion2, so we make sure to not generate those queries in combination with ShelleyNodeToClientVersion1.

Instance details

Methods

arbitrary ∷ Gen (WithVersion ShelleyNodeToClientVersion (SomeSecond BlockQuery (ShelleyBlock era)))

shrink ∷ WithVersion ShelleyNodeToClientVersion (SomeSecond BlockQuery (ShelleyBlock era)) → [WithVersion ShelleyNodeToClientVersion (SomeSecond BlockQuery (ShelleyBlock era))]

Arbitrary (PParams' StrictMaybe era)

Instance details

Methods

arbitrary ∷ Gen (PParams' StrictMaybe era)

shrink ∷ PParams' StrictMaybe era → [PParams' StrictMaybe era]