Test.Consensus.Shelley.Generators

Documentation

data SomeResult where

To easily generate all the possible results of the Query GADT, we introduce an existential that also bundles the corresponding Query as evidence. We also capture an Eq and a Show constraint, as we need them in the tests.

Constructors

SomeResult ∷ (Eq result, Show result) ⇒ Query Block result → result → SomeResult

genHash ∷ ∀ a c. Crypto c ⇒ Proxy c → Gen (Hash c a)

Orphan instances

Arbitrary Natural

Instance details

Methods

arbitrary ∷ Gen Natural

shrink ∷ Natural → [Natural]

Arbitrary ProtVer

Instance details

Methods

arbitrary ∷ Gen ProtVer

shrink ∷ ProtVer → [ProtVer]

Arbitrary IPv4

Instance details

Methods

arbitrary ∷ Gen IPv4

shrink ∷ IPv4 → [IPv4]

Arbitrary AccountState

Instance details

Methods

arbitrary ∷ Gen AccountState

shrink ∷ AccountState → [AccountState]

Arbitrary ApparentPerformance

Instance details

Methods

arbitrary ∷ Gen ApparentPerformance

shrink ∷ ApparentPerformance → [ApparentPerformance]

Arbitrary PoolMetaData

Instance details

Methods

arbitrary ∷ Gen PoolMetaData

shrink ∷ PoolMetaData → [PoolMetaData]

Arbitrary StakePoolRelay

Instance details

Methods

arbitrary ∷ Gen StakePoolRelay

shrink ∷ StakePoolRelay → [StakePoolRelay]

Arbitrary Ptr

Instance details

Methods

arbitrary ∷ Gen Ptr

shrink ∷ Ptr → [Ptr]

Arbitrary PParams

Instance details

Methods

arbitrary ∷ Gen PParams

shrink ∷ PParams → [PParams]

Arbitrary UnitInterval

Instance details

Methods

arbitrary ∷ Gen UnitInterval

shrink ∷ UnitInterval → [UnitInterval]

Arbitrary Nonce

Instance details

Methods

arbitrary ∷ Gen Nonce

shrink ∷ Nonce → [Nonce]

Arbitrary Url

Instance details

Methods

arbitrary ∷ Gen Url

shrink ∷ Url → [Url]

Arbitrary DnsName

Instance details

Methods

arbitrary ∷ Gen DnsName

shrink ∷ DnsName → [DnsName]

Arbitrary Port

Instance details

Methods

arbitrary ∷ Gen Port

shrink ∷ Port → [Port]

Arbitrary ActiveSlotCoeff

Instance details

Methods

arbitrary ∷ Gen ActiveSlotCoeff

shrink ∷ ActiveSlotCoeff → [ActiveSlotCoeff]

Arbitrary Coin

Instance details

Methods

arbitrary ∷ Gen Coin

shrink ∷ Coin → [Coin]

Arbitrary IPv6

Instance details

Methods

arbitrary ∷ Gen IPv6

shrink ∷ IPv6 → [IPv6]

Arbitrary Block

Instance details

Methods

arbitrary ∷ Gen Block

shrink ∷ Block → [Block]

Arbitrary (PredicateFailure (LEDGERS TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (LEDGERS TPraosMockCrypto))

shrink ∷ PredicateFailure (LEDGERS TPraosMockCrypto) → [PredicateFailure (LEDGERS TPraosMockCrypto)]

Arbitrary (PredicateFailure (LEDGER TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (LEDGER TPraosMockCrypto))

shrink ∷ PredicateFailure (LEDGER TPraosMockCrypto) → [PredicateFailure (LEDGER TPraosMockCrypto)]

Arbitrary (PredicateFailure (DELEGS TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (DELEGS TPraosMockCrypto))

shrink ∷ PredicateFailure (DELEGS TPraosMockCrypto) → [PredicateFailure (DELEGS TPraosMockCrypto)]

Arbitrary (PredicateFailure (DELPL TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (DELPL TPraosMockCrypto))

shrink ∷ PredicateFailure (DELPL TPraosMockCrypto) → [PredicateFailure (DELPL TPraosMockCrypto)]

Arbitrary (PredicateFailure (DELEG TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (DELEG TPraosMockCrypto))

shrink ∷ PredicateFailure (DELEG TPraosMockCrypto) → [PredicateFailure (DELEG TPraosMockCrypto)]

Arbitrary (PredicateFailure (POOL TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (POOL TPraosMockCrypto))

shrink ∷ PredicateFailure (POOL TPraosMockCrypto) → [PredicateFailure (POOL TPraosMockCrypto)]

Arbitrary (PredicateFailure (UTXOW TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (UTXOW TPraosMockCrypto))

shrink ∷ PredicateFailure (UTXOW TPraosMockCrypto) → [PredicateFailure (UTXOW TPraosMockCrypto)]

Arbitrary (PredicateFailure (UTXO TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (UTXO TPraosMockCrypto))

shrink ∷ PredicateFailure (UTXO TPraosMockCrypto) → [PredicateFailure (UTXO TPraosMockCrypto)]

Arbitrary (PredicateFailure (PPUP TPraosMockCrypto))

Instance details

Methods

arbitrary ∷ Gen (PredicateFailure (PPUP TPraosMockCrypto))

shrink ∷ PredicateFailure (PPUP TPraosMockCrypto) → [PredicateFailure (PPUP TPraosMockCrypto)]

Arbitrary a ⇒ Arbitrary (StrictSeq a)

Instance details

Methods

arbitrary ∷ Gen (StrictSeq a)

shrink ∷ StrictSeq a → [StrictSeq a]

Arbitrary (Point Block)

Instance details

Methods

arbitrary ∷ Gen (Point Block)

shrink ∷ Point Block → [Point Block]

Arbitrary (GenTx Block)

Instance details

Methods

arbitrary ∷ Gen (GenTx Block)

shrink ∷ GenTx Block → [GenTx Block]

Arbitrary (GenTxId Block)

Instance details

Methods

arbitrary ∷ Gen (GenTxId Block)

shrink ∷ GenTxId Block → [GenTxId Block]

Arbitrary (LedgerState Block)

Instance details

Methods

arbitrary ∷ Gen (LedgerState Block)

shrink ∷ LedgerState Block → [LedgerState Block]

Arbitrary (Header Block)

Instance details

Methods

arbitrary ∷ Gen (Header Block)

shrink ∷ Header Block → [Header Block]

Arbitrary (TxIn TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (TxIn TPraosMockCrypto)

shrink ∷ TxIn TPraosMockCrypto → [TxIn TPraosMockCrypto]

Arbitrary (NonMyopicMemberRewards TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (NonMyopicMemberRewards TPraosMockCrypto)

shrink ∷ NonMyopicMemberRewards TPraosMockCrypto → [NonMyopicMemberRewards TPraosMockCrypto]

Arbitrary (LedgerViewHistory TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (LedgerViewHistory TPraosMockCrypto)

shrink ∷ LedgerViewHistory TPraosMockCrypto → [LedgerViewHistory TPraosMockCrypto]

Arbitrary (TPraosState TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (TPraosState TPraosMockCrypto)

shrink ∷ TPraosState TPraosMockCrypto → [TPraosState TPraosMockCrypto]

Crypto c ⇒ Arbitrary (ShelleyHash c)

Instance details

Methods

arbitrary ∷ Gen (ShelleyHash c)

shrink ∷ ShelleyHash c → [ShelleyHash c]

Crypto c ⇒ Arbitrary (LedgerView c)

Instance details

Methods

arbitrary ∷ Gen (LedgerView c)

shrink ∷ LedgerView c → [LedgerView c]

Arbitrary (ApplyTxError TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (ApplyTxError TPraosMockCrypto)

shrink ∷ ApplyTxError TPraosMockCrypto → [ApplyTxError TPraosMockCrypto]

Crypto c ⇒ Arbitrary (PrtclState c)

Instance details

Methods

arbitrary ∷ Gen (PrtclState c)

shrink ∷ PrtclState c → [PrtclState c]

Crypto c ⇒ Arbitrary (RewardUpdate c)

Instance details

Methods

arbitrary ∷ Gen (RewardUpdate c)

shrink ∷ RewardUpdate c → [RewardUpdate c]

Arbitrary (EpochState TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (EpochState TPraosMockCrypto)

shrink ∷ EpochState TPraosMockCrypto → [EpochState TPraosMockCrypto]

Crypto c ⇒ Arbitrary (OBftSlot c)

Instance details

Methods

arbitrary ∷ Gen (OBftSlot c)

shrink ∷ OBftSlot c → [OBftSlot c]

Arbitrary (NewEpochState TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (NewEpochState TPraosMockCrypto)

shrink ∷ NewEpochState TPraosMockCrypto → [NewEpochState TPraosMockCrypto]

Arbitrary (LedgerState TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (LedgerState TPraosMockCrypto)

shrink ∷ LedgerState TPraosMockCrypto → [LedgerState TPraosMockCrypto]

Arbitrary (NonMyopic TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (NonMyopic TPraosMockCrypto)

shrink ∷ NonMyopic TPraosMockCrypto → [NonMyopic TPraosMockCrypto]

Crypto c ⇒ Arbitrary (BlocksMade c)

Instance details

Methods

arbitrary ∷ Gen (BlocksMade c)

shrink ∷ BlocksMade c → [BlocksMade c]

Crypto c ⇒ Arbitrary (Stake c)

Instance details

Methods

arbitrary ∷ Gen (Stake c)

shrink ∷ Stake c → [Stake c]

Arbitrary (SnapShot TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (SnapShot TPraosMockCrypto)

shrink ∷ SnapShot TPraosMockCrypto → [SnapShot TPraosMockCrypto]

Arbitrary (SnapShots TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (SnapShots TPraosMockCrypto)

shrink ∷ SnapShots TPraosMockCrypto → [SnapShots TPraosMockCrypto]

Crypto c ⇒ Arbitrary (PoolDistr c)

Instance details

Methods

arbitrary ∷ Gen (PoolDistr c)

shrink ∷ PoolDistr c → [PoolDistr c]

Arbitrary (Tx TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (Tx TPraosMockCrypto)

shrink ∷ Tx TPraosMockCrypto → [Tx TPraosMockCrypto]

Crypto c ⇒ Arbitrary (PoolParams c)

Instance details

Methods

arbitrary ∷ Gen (PoolParams c)

shrink ∷ PoolParams c → [PoolParams c]

Arbitrary (TxOut TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (TxOut TPraosMockCrypto)

shrink ∷ TxOut TPraosMockCrypto → [TxOut TPraosMockCrypto]

Crypto c ⇒ Arbitrary (Addr c)

Instance details

Methods

arbitrary ∷ Gen (Addr c)

shrink ∷ Addr c → [Addr c]

Crypto c ⇒ Arbitrary (StakeReference c)

Instance details

Methods

arbitrary ∷ Gen (StakeReference c)

shrink ∷ StakeReference c → [StakeReference c]

Crypto c ⇒ Arbitrary (RewardAcnt c)

Instance details

Methods

arbitrary ∷ Gen (RewardAcnt c)

shrink ∷ RewardAcnt c → [RewardAcnt c]

Arbitrary (PParams' StrictMaybe)

Instance details

Methods

arbitrary ∷ Gen (PParams' StrictMaybe)

shrink ∷ PParams' StrictMaybe → [PParams' StrictMaybe]

Arbitrary (ProposedPPUpdates TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (ProposedPPUpdates TPraosMockCrypto)

shrink ∷ ProposedPPUpdates TPraosMockCrypto → [ProposedPPUpdates TPraosMockCrypto]

Crypto c ⇒ Arbitrary (MetaDataHash c)

Instance details

Methods

arbitrary ∷ Gen (MetaDataHash c)

shrink ∷ MetaDataHash c → [MetaDataHash c]

Crypto c ⇒ Arbitrary (ScriptHash c)

Instance details

Methods

arbitrary ∷ Gen (ScriptHash c)

shrink ∷ ScriptHash c → [ScriptHash c]

Crypto c ⇒ Arbitrary (GenDelegs c)

Instance details

Methods

arbitrary ∷ Gen (GenDelegs c)

shrink ∷ GenDelegs c → [GenDelegs c]

Arbitrary a ⇒ Arbitrary (StrictMaybe a)

Instance details

Methods

arbitrary ∷ Gen (StrictMaybe a)

shrink ∷ StrictMaybe a → [StrictMaybe a]

Eq (Some (Query Block))

Instance details

Methods

(==) ∷ Some (Query Block) → Some (Query Block) → Bool

(/=) ∷ Some (Query Block) → Some (Query Block) → Bool

Show (Some (Query Block))

Instance details

Methods

showsPrec ∷ Int → Some (Query Block) → ShowS

show ∷ Some (Query Block) → String

showList ∷ [Some (Query Block)] → ShowS

Arbitrary (Some (Query Block))

Instance details

Methods

arbitrary ∷ Gen (Some (Query Block))

shrink ∷ Some (Query Block) → [Some (Query Block)]

Crypto c ⇒ Arbitrary (Credential r c)

Instance details

Methods

arbitrary ∷ Gen (Credential r c)

shrink ∷ Credential r c → [Credential r c]

Arbitrary (VKey kd TPraosMockCrypto)

Instance details

Methods

arbitrary ∷ Gen (VKey kd TPraosMockCrypto)

shrink ∷ VKey kd TPraosMockCrypto → [VKey kd TPraosMockCrypto]

Crypto c ⇒ Arbitrary (KeyHash a c)

Instance details

Methods

arbitrary ∷ Gen (KeyHash a c)

shrink ∷ KeyHash a c → [KeyHash a c]