Test.Consensus.Shelley.MockCrypto

Documentation

data MockCrypto h

A mock replacement for StandardCrypto

We run the tests with this mock crypto, as it is easier to generate and debug things. The code is parametric in the crypto, so it shouldn't make much of a difference. This also has the important advantage that we can reuse the generators from cardano-ledger-specs.

type MockShelley h = Shelley (MockCrypto h)

type Block h = ShelleyBlock (MockShelley h)

type CanMock era = (Era era, Mock (Crypto era))

Cryptography that can easily be mocked