Test.Consensus.Shelley.MockCrypto

Documentation

type Block h = ShelleyBlock (MockShelley h)

type CanMock era = (ShelleyBasedEra era, EraGen era, Mock (EraCrypto era), ValidateScript era, Arbitrary (AuxiliaryData era), Arbitrary (PParams era), Arbitrary (Script era), Arbitrary (TxBody era), Arbitrary (Tx era), Arbitrary (TxOut era), Arbitrary (Value era), Arbitrary (PredicateFailure (UTXOW era)), Arbitrary (Witnesses era))

Cryptography that can easily be mocked

data MockCrypto h

A mock replacement for StandardCrypto

We run the tests with this mock crypto, as it is easier to generate and debug things. The code is parametric in the crypto, so it shouldn't make much of a difference. This also has the important advantage that we can reuse the generators from cardano-ledger-specs.

type MockShelley h = ShelleyEra (MockCrypto h)