Test.Consensus.Shelley.MockCrypto

Documentation

type TPraosMockCrypto = ConcreteCrypto

Use ConcreteCrypto from cardano-ledger-specs as our mock crypto. This means we can reuse all their examples and generators.

We call it TPraosMockCrypto, because the name ConcreteCrypto is confusing, "concrete" gives the impression it is the real crypto, but it's not (that's our TPraosStandardCrypto).

type Block = ShelleyBlock TPraosMockCrypto

We run the tests with mock crypto, as it is easier to generate and debug things. The code is parametric in the crypto, so it shouldn't make much of a difference. This also has the important advantage that we can reuse the generators from cardano-ledger-specs.

Orphan instances

TPraosCrypto TPraosMockCrypto

Instance details