Test.ThreadNet.Infra.Shelley

Documentation

data CoreNode c

Constructors

CoreNode

Fields

cnGenesisKey ∷ !(SignKeyDSIGN c)

cnDelegateKey ∷ !(SignKeyDSIGN c)

Cold delegate key. The hash of the corresponding verification (public) key will be used as the payment credential.

cnStakingKey ∷ !(SignKeyDSIGN c)

The hash of the corresponding verification (public) key will be used as the staking credential.

cnVRF ∷ !(SignKeyVRF c)

cnKES ∷ !(SignKeyKES c)

cnOCert ∷ !(OCert c)

data CoreNodeKeyInfo c

Constructors

CoreNodeKeyInfo

Fields

cnkiKeyPair ∷ (KeyPair 'Payment c, KeyPair 'Staking c)

cnkiCoreNode ∷ (KeyPair 'Genesis c, AllIssuerKeys c 'GenesisDelegate)

newtype DecentralizationParam

A suitable value for the d protocol parameter

In the range 0 to 1, inclusive. Beware the misnomer: 0 means fully decentralized, and 1 means fully centralized.

Constructors

DecentralizationParam

Fields

decentralizationParamToRational ∷ Rational

data KesConfig

Currently maxEvolutions * slotsPerEvolution is the max number of slots the test can run without needing new ocerts.

TODO This limitation may be lifted by PR #2107, see https://github.com/input-output-hk/ouroboros-network/issues/2107.

Constructors

KesConfig

Fields

maxEvolutions ∷ Word64

slotsPerEvolution ∷ Word64

coreNodeKeys ∷ ∀ c. PraosCrypto c ⇒ CoreNode c → CoreNodeKeyInfo c

genCoreNode ∷ ∀ c. PraosCrypto c ⇒ KESPeriod → Gen (CoreNode c)

incrementMinorProtVer ∷ ProtVer → ProtVer

initialLovelacePerCoreNode ∷ Word64

mkCredential ∷ PraosCrypto c ⇒ SignKeyDSIGN c → Credential r c

mkEpochSize ∷ SecurityParam → Rational → EpochSize

The epoch size, given k and f.

INVARIANT: 10 * k / f must be a whole number.

mkGenesisConfig

Arguments

∷ ∀ era. PraosCrypto (EraCrypto era)

⇒ ProtVer
Initial protocol version

→ SecurityParam

→ Rational
Initial active slot coefficient

→ DecentralizationParam

→ Word64
Max Lovelace supply, must be >= #coreNodes * initialLovelacePerCoreNode

→ SlotLength

→ KesConfig

→ [CoreNode (EraCrypto era)]

→ ShelleyGenesis era

Note: a KES algorithm supports a particular max number of KES evolutions, but we can configure a potentially lower maximum for the ledger, that's why we take it as an argument.

mkKesConfig ∷ ∀ proxy c. Crypto c ⇒ proxy c → NumSlots → KesConfig

A KesConfig that will not require more evolutions than this test's crypto allows.

mkKeyHash ∷ PraosCrypto c ⇒ SignKeyDSIGN c → KeyHash r c

mkKeyHashVrf ∷ (HashAlgorithm h, VRFAlgorithm vrf) ⇒ SignKeyVRF vrf → Hash h (VerKeyVRF vrf)

mkKeyPair ∷ PraosCrypto c ⇒ SignKeyDSIGN c → KeyPair r c

mkLeaderCredentials ∷ PraosCrypto c ⇒ CoreNode c → TPraosLeaderCredentials c

mkMASetDecentralizationParamTxs

Arguments

∷ ∀ era. (ShelleyBasedEra era, Tx era ~ Tx era, TxBody era ~ TxBody era, PParams era ~ PParams era, PParamsDelta era ~ PParams' StrictMaybe era, Witnesses era ~ WitnessSet era)

⇒ [CoreNode (Crypto era)]

→ ProtVer
The proposed protocol version

→ SlotNo
The TTL

→ DecentralizationParam
The new value

→ [GenTx (ShelleyBlock era)]

TODO This is a copy-paste-edit of mkSetDecentralizationParamTxs

Our current plan is to replace all of this infrastructure with the ThreadNet rewrite; so we're minimizing the work and maintenance here for now.

mkProtocolShelley ∷ ∀ m c. (IOLike m, ShelleyBasedEra (ShelleyEra c)) ⇒ ShelleyGenesis (ShelleyEra c) → Nonce → ProtVer → CoreNode c → ProtocolInfo m (ShelleyBlock (ShelleyEra c))

mkSetDecentralizationParamTxs

Arguments

∷ ∀ c. ShelleyBasedEra (ShelleyEra c)

⇒ [CoreNode c]

→ ProtVer
The proposed protocol version

→ SlotNo
The TTL

→ DecentralizationParam
The new value

→ [GenTx (ShelleyBlock (ShelleyEra c))]

mkVerKey ∷ PraosCrypto c ⇒ SignKeyDSIGN c → VKey r c

networkId ∷ Network

tpraosSlotLength ∷ SlotLength